% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af1_clim.R
\docType{data}
\name{af1_clim}
\alias{af1_clim}
\title{Monthly climate data for Ancient Forest site in the ICH}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 14 columns.
}
\usage{
af1_clim
}
\description{
Monthly climate data for Ancient Forest site in the ICH
}
\details{
{

Example data for use in testing functions.  Climate data generated using ClimateBC v7.2

Two climate variables: Precipitation (PPT) and mean maximum daily temperatures (Tmax).

}
}
\keyword{datasets}
