% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af1_crn.R
\docType{data}
\name{af1_crn}
\alias{af1_crn}
\title{Western redcedar ring-width chronology from Ancient Forest}
\format{
An object of class \code{data.frame} with 264 rows and 1 columns.
}
\usage{
af1_crn
}
\description{
Western redcedar ring-width chronology from Ancient Forest
}
\details{
{

Example data for use in testing functions.  Chronology is a residual chronology from cedar cores collected
in 2021.

Year is in rownames.  You can bring in into the dataframe using the \link[tibble]{rownames_to_column} function.
}
}
\keyword{datasets}
