% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_boot.R
\name{corr_boot}
\alias{corr_boot}
\title{Bootstrapped correlation coefficients}
\usage{
corr_boot(dat, x, y, nboot = 1000)
}
\arguments{
\item{dat}{data frame}

\item{x}{name of first column (function will fail if not a character).}

\item{y}{name of second column (function will fail if not a character).}

\item{nboot}{Number of replicates for bootstrapping.  Defaults to 1000.}
}
\value{
A data frame with five columns:
\itemize{
\item{corr_coef: correlation coefficient, using cor.test}
\item{p_value: p value of the test, using cor.test}
\item{n: number of observations used to calculate coefficient}
\item{LCL: Lower 95\% confidence interval}
\item{UCL: Upper 95\% confidence interval}
\item{SIG: Significance based on CI}

}
}
\description{
This function simply returns a correlation coefficient (and other statistics) between
two variables.  Function uses the cor.test and boot.ci functions.
}
\details{
{
\itemize{
\item Confidence intervals are 95\%.
\item Bootstrapped CI are bias-corrected accelerated.
\item NA values are allowed.
}

}
}
\examples{

# Join climate and residual chronology data for Ancient Forest together
 af1_clim \%>\%
 dplyr::filter(var=="Tmax") \%>\%
 dplyr::left_join(af1_crn \%>\%
             tibble::rownames_to_column("year") \%>\%
             dplyr::mutate(year=as.integer(year))) \%>\%
 corr_boot(x="res",y="Jan")



}
