% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_dbh_class.R
\name{cut_dbh_class}
\alias{cut_dbh_class}
\title{Cut diameters into classes}
\usage{
cut_dbh_class(x, size_class = 5, min_size = 7.5, labels = "mid")
}
\arguments{
\item{x}{Vector of diameter data.}

\item{size_class}{Width of individual size classes.  Defaults to 5 (cm).}

\item{min_size}{Lowest value of first interval. Defaults to 7.5 (cm).}

\item{labels}{Must be one of c("mid","min","max","label").  Specify what type of cut labels you want.  Defaults to mid-point.}
}
\value{
A vector with labels.
}
\description{
This function cuts a vector of diameter measurements into discrete diameter classes.  User specifies diameter class width and type of label desired.
}
\details{
Note that cuts are not closed on the right.  This means that a particular value falling on a cut boundary will belong to the next largest class, not smallest.
For example, a tree with 10.0cm will be placed in the (10.0-15.0) class, not the (5.0-10.0) class.
}
\examples{
cut_dbh_class(x=runif(min=7.5,max=56.7,n=100),labels="min")

data.frame(dbh=runif(min=7.5,max=56.7,n=100)) \%>\%
dplyr::mutate(dbh_class_mid=cut_dbh_class(dbh),
      dbh_class_label=cut_dbh_class(x=dbh,labels="label"))
}
