% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2ba.R
\name{d2ba}
\alias{d2ba}
\title{Calculate tree basal area from diameter}
\usage{
d2ba(x)
}
\arguments{
\item{x}{vector of diameter data.}
}
\value{
A vector with basal area.
}
\description{
Simple function to convert tree diameter to basal area.
}
\details{
Function assumes that diameter is measured in centimetres.  Basal area is returned in metres.
}
\examples{
data.frame(dbh=runif(min=7.5,max=56.7,n=100)) \%>\%
dplyr::mutate(ba=d2ba(dbh))
}
