% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_p_values.R
\name{print_p_values}
\alias{print_p_values}
\title{Symbolize p values}
\usage{
print_p_values(
  x,
  sig_levels = c(0.001, 0.01, 0.049999),
  symbol_levels = c("***", "**", "*", " ")
)
}
\arguments{
\item{x}{vector of p values.}

\item{sig_levels}{optional. Cutoff points for p values.}

\item{symbol_levels}{optional. Different symbols to represent p values}
}
\description{
Simple function to convert p values to asterisks for printing.
}
\examples{
 data.frame(p_values=c(0.05,0.023,0.0009,0.0001,0.003)) \%>\%
dplyr::mutate(p_print=print_p_values(p_values))
}
