% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_coords.R
\name{tree_coords}
\alias{tree_coords}
\title{Tree spatial coordinates}
\usage{
tree_coords(data, azi, dist)
}
\arguments{
\item{data}{data must be provided in data frame.}

\item{azi}{name of column in 'data' that contains bearing from plot centre to trees, in degrees.}

\item{dist}{name of column in 'data' that contains distance between trees and plot centre, in metres.}
}
\value{
A data frame with two columns (x and y cartesian coordinates) and same rows as input data.
}
\description{
Convert stem mapping data (azimuth and distance from plot centre) to xy coordinates
}
\details{
This function simply takes the azimuth and distance between plot centre and individual tree positions, and converts them to xy coordinates to
allow for stem mapping and neighborhood analysis.

Note: this function makes a number of assumptions:
\itemize{
\item{Azimuth}{ is recorded from plot centre to individual tree position, and is recorded in
true north, not magnetic north direction (In other words, function assumes that the data are corrected for declination); and}
\item{Distance}{ is horizontal distance, in metres.  No correction is made for slope (you must do this separately).}
}
}
